requirements.txt
msoffcrypto-tool>=5.0.0
openpyxl>=3.1.0
